program working for float and -ve nos

+++ src/sandeshm/expeval/lib/expressionEvaluator.java
@@ -1,8 +1,9 @@
 package sandeshm.expeval.lib;
 
 public class expressionEvaluator {
-    public int evaluateExpression(String expression) throws Exception {
+    public double evaluateExpression(String expression) throws Exception {
         int isParanthesisInExpr = expression.indexOf('(');
+
         String[] splittedExpression = expression.split(" ");
         Operations operate = new Operations();
